Create book component
