+++ src/Components/Boolist/Book.js
@@ -1,42 +1,19 @@
-import '../Style/Style.css';
-import NewBookForm from './NewBookForm';
+import propTypes from 'prop-types';
 
-const Book = () => {
-  const data = [
-    {
-      id: 1,
-      title: 'Setup development environment',
-      author: ' Okpara Blessing',
-    },
-    {
-      id: 2,
-      title: 'Develop website and add content',
-      author: ' Okpara',
-    },
-    {
-      id: 3,
-      title: 'Deploy to live server',
-      author: ' Blessing',
-    },
-  ];
-  return (
-    <>
+const MyBook = ({ title, author }) => (
+  <div className="eachBook">
+    <h2>{title}</h2>
+    <p>
+      By
+      {author}
+    </p>
+    <button type="button" id="remove">Remove</button>
+  </div>
+);
 
-      <div className="container">
-        {data.map((item) => (
-          <div className="eachBook" key={item.id}>
-            <h2>{item.title}</h2>
-            <p>
-              By
-              {item.author}
-            </p>
-            <button type="button" id="remove">Remove</button>
-          </div>
-        ))}
-      </div>
-      <NewBookForm />
-    </>
-  );
+MyBook.propTypes = {
+  title: propTypes.string.isRequired,
+  author: propTypes.string.isRequired,
 };
 
-export default Book;
+export default MyBook;
